20000426-00:22 GMT+1 Victor Szakats <info@szelvesz.hu>

+++ harbour/source/rtl/transfrm.c
@@ -368,7 +368,7 @@ HB_FUNC( TRANSFORM )
          if( uiPicFlags & PF_STRING )
          {
             char * szPic2 = hb_itemGetCPtr( pPic );
-            int ulLen = atol( szPic2 + 2 );
+            ULONG ulLen = ( ULONG ) atol( szPic2 + 2 );
 
             if( ulLen & ( ulLen <= ulExpLen ) )
             {
